Scent is now between 0 and 1, and higher scent decreases pheromone strength

+++ Part1/cell.h
@@ -23,7 +23,7 @@ public:
     virtual void reset();
 
     bool hasPheremone() const;
-    int getPheremone();
+    double getPheremone();
 
     void setAnthill();
     void setFood(QSP<Food> food);
